refactor(mode): move lifecycle into src/session, webview pointer into input/default_mode
