+++ src/mode.rs
@@ -1,4 +1,3 @@
 //! Per-mode processing; each mode declares its systems inside a dedicated submodule.
 
-pub(crate) mod default;
 pub(crate) mod tmux;
